chore: move package to net.mamoe.mirai.mock.internal.absoluteFile
