+++ mirai-core-mock/src/internal/absolutefile/MockAbsoluteFolder.kt
@@ -7,7 +7,7 @@
  * https://github.com/mamoe/mirai/blob/dev/LICENSE
  */
 
-package net.mamoe.mirai.mock.internal.remotefile.v2
+package net.mamoe.mirai.mock.internal.absolutefile
 
 import kotlinx.coroutines.flow.*
 import net.mamoe.mirai.contact.FileSupported
