+++ mirai-core-mock/src/internal/absolutefile/utils.kt
@@ -7,7 +7,7 @@
  * https://github.com/mamoe/mirai/blob/dev/LICENSE
  */
 
-package net.mamoe.mirai.mock.internal.remotefile.v2
+package net.mamoe.mirai.mock.internal.absolutefile
 
 import net.mamoe.mirai.contact.file.AbsoluteFile
 import net.mamoe.mirai.contact.file.AbsoluteFolder
